UPDATE ALL + Traduction
Création du check du nb de travaux +
Vérification du menu + déplacement de celui ci dans _To5
Mise en place de traduction dans le code, pas fini
J'ai supprimé 2 classes, car useless (Enum/software)
Reste à faire : Fichier unique

+++ Livrable/WriteLog.cs
@@ -3,28 +3,30 @@ using System.IO;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Newtonsoft.Json;
+using System.Collections.Generic;
 
 
 namespace Livrable
 {
     public class WriteLog
     {
-        public Save sauvegarde;//CREER LA SAUVEGARDE EN ATTENDANT QUE LUILISATEUR LE FASSE
+        public Save save;
         public string source;
         public string dest;
         public double taille;
         public CalculLenght calculTaille = new CalculLenght();
         public CalculTime calculTemps;
         public double temps;
+        List<JsonTry> listJSON = new List<JsonTry>();
 
 
         public WriteLog(string fichierSource, string fichierDest)//construct a log with a source, a destination, a length and a time
         {
-            this.sauvegarde = new Save("Test", fichierSource, fichierDest);
+            this.save = new Save("Test", fichierSource, fichierDest);
             this.source = fichierSource;
             this.dest = fichierDest;
-            this.taille = this.sauvegarde.tailleFichiers;
-            calculTemps = new CalculTime(this.sauvegarde);
+            this.taille = this.save.tailleFichiers;
+            calculTemps = new CalculTime(this.save);
             this.temps = calculTemps.temps;
         }
 
@@ -35,7 +37,7 @@ namespace Livrable
 
                 JsonTry log = new JsonTry();
                 log.Date = Save.horodatage;
-                log.name = sauvegarde.appellation;
+                log.name = save.appellation;
                 log.dest = dest;
                 log.source = source;
                 log.Lenght = taille;
@@ -43,15 +45,25 @@ namespace Livrable
                 string json = JsonConvert.SerializeObject(log);
 
                 string fileName = @"/Users/aymerick/Desktop/CESI/test.json";
-                using (StreamWriter writer = new StreamWriter(@fileName, true))
+
+
+                if (!File.Exists(fileName))
                 {
-                    
+                    listJSON.Add(log);
                     var options = new JsonSerializerOptions { WriteIndented = true };
-                    string jsonString = System.Text.Json.JsonSerializer.Serialize(log, options);
-                    Console.Write("--- Affichage du log ----\n");
-                    Console.Write(jsonString);
-                    Console.Write("\n--------\n");
+                    string jsonString = System.Text.Json.JsonSerializer.Serialize(listJSON, options);
                     File.WriteAllText(@fileName, jsonString);
+                }
+                else
+                {
+                    StreamReader r = new StreamReader(fileName);
+                    string jsonString2 = r.ReadToEnd();
+                    r.Close();
+                    List<JsonTry> m = JsonConvert.DeserializeObject<List<JsonTry>>(jsonString2);
+                    m.Add(log);
+                    var options = new JsonSerializerOptions { WriteIndented = true };
+                    string jsonString3 = System.Text.Json.JsonSerializer.Serialize(m, options);
+                    File.WriteAllText(@fileName, jsonString3);
                 }
             }
             catch (Exception exp)
@@ -62,3 +74,4 @@ namespace Livrable
         }
     }
 }
+
